Revert "Isolate UIKit-specific Plume APIs"
This reverts commit 11cc89766d7da95c67141317c4e6e17a849cdf8e.
