+++ Sources/Plume/Public/Extensions/Array/Array+Extensions.swift
@@ -5,12 +5,55 @@
 //  Created by Samuel Lupton on 4/26/26.
 //
 
-import CoreGraphics
+#if canImport(UIKit)
+import UIKit
+#else
+import Foundation
+#endif
 
 // MARK: - Plume Cell Array Factories
 
 /// Convenience factories for building arrays of `Plume.Cell` from image collections.
 extension Array where Element == Plume.Cell {
+
+    #if canImport(UIKit)
+    /// Creates an array of `Plume.Cell` from a collection of `UIImage` values.
+    ///
+    /// Each image is converted into a particle while sharing a common set of
+    /// behavioral properties such as lifetime, velocity, and spin. This provides
+    /// a concise way to generate multiple cells with consistent behavior.
+    ///
+    /// - Parameters:
+    ///   - uiimages: The images used as the visual contents of each cell.
+    ///   - lifetime: The lifetime configuration applied to all cells.
+    ///   - spin: The rotational behavior applied to all cells.
+    ///   - scale: The size configuration applied to all cells.
+    ///   - acceleration: The acceleration applied to all cells.
+    ///   - velocity: The velocity applied to all cells.
+    ///   - angle: The emission angle applied to all cells.
+    /// - Returns: An array of configured `Plume.Cell` instances.
+    public static func make(
+        from uiimages: [UIImage],
+        lifetime: Plume.Cell.Lifetime,
+        spin: Plume.Cell.Spin,
+        scale: Plume.Cell.Scale,
+        acceleration: Plume.Cell.Acceleration,
+        velocity: Plume.Cell.Velocity,
+        angle: Plume.Cell.Angle
+    ) -> [Plume.Cell] {
+        uiimages.map { uiimage in
+            Plume.Cell(
+                contents: Plume.Cell.Contents(uiimage: uiimage),
+                lifetime: lifetime,
+                spin: spin,
+                scale: scale,
+                acceleration: acceleration,
+                velocity: velocity,
+                angle: angle
+            )
+        }
+    }
+    #endif
 
     /// Creates an array of `Plume.Cell` from a collection of `CGImage` values.
     ///
@@ -48,4 +91,40 @@ extension Array where Element == Plume.Cell {
         }
     }
 
+    /// Creates an array of `Plume.Cell` from a collection of `ImageResource` values.
+    ///
+    /// This variant integrates with Swift’s resource system (iOS 17+),
+    /// allowing asset-based particle creation with the same shared behavior.
+    ///
+    /// - Parameters:
+    ///   - resources: The image resources used as the visual contents of each cell.
+    ///   - lifetime: The lifetime configuration applied to all cells.
+    ///   - spin: The rotational behavior applied to all cells.
+    ///   - scale: The size configuration applied to all cells.
+    ///   - acceleration: The acceleration applied to all cells.
+    ///   - velocity: The velocity applied to all cells.
+    ///   - angle: The emission angle applied to all cells.
+    /// - Returns: An array of configured `Plume.Cell` instances.
+    @available(iOS 17.0, *)
+    public static func make(
+        from resources: [ImageResource],
+        lifetime: Plume.Cell.Lifetime,
+        spin: Plume.Cell.Spin,
+        scale: Plume.Cell.Scale,
+        acceleration: Plume.Cell.Acceleration,
+        velocity: Plume.Cell.Velocity,
+        angle: Plume.Cell.Angle
+    ) -> [Plume.Cell] {
+        resources.map { resource in
+            Plume.Cell(
+                contents: Plume.Cell.Contents(resource: resource),
+                lifetime: lifetime,
+                spin: spin,
+                scale: scale,
+                acceleration: acceleration,
+                velocity: velocity,
+                angle: angle
+            )
+        }
+    }
 }
