+++ Sources/Plume/Public/UI/PlumeUIView.swift
@@ -7,6 +7,9 @@
 
 #if canImport(UIKit)
 import UIKit
+#else
+import Foundation
+#endif
 
 // TODO: Consider if this should be a public facing component.
 // TODO: Add support for haptics.
@@ -133,4 +136,3 @@ extension PlumeUIView {
         }
     }
 }
-#endif
